dana_service_SOA/requirements.txt
python-dotenv
PyMySQL
spyne
lxml
zeep
requests
PyJWT